Title: 10개 입력받기, Time: 242ms, Memory: 15MB, Status: Wrong Answer - CodeTree

+++ 241001/10개 입력받기/receive-10-inputs.java
@@ -14,7 +14,7 @@ public class Main {
                 }
             }
             int sum = Arrays.stream(arr).sum();
-            long cnt = Arrays.stream(arr)
+            double cnt = Arrays.stream(arr)
                         .filter(num -> num != 0)
                         .count();
             double avg = sum / cnt;
